ddtrace/tracer: fix remote config memory leak
Prior to this change, we'd always keep around entries in the
map corresponding to zero-byte updates. These are sent for
configurations that have been removed. This amounts to a memory
leak: the map will contain every key the process has ever seen.

In practice nobody uses this feature really, and when they do
they don't tend to produce tons and tons of probes, but it was
also annoying seeing these empty calls on the consuming side.