chore: move TestIntegrationEnabled from pkg ddtrace/tracer to instrumentation
The main motivation is that this test is very slow (~35s on my machine)
and has a big impact on the feedback loop of working on the
ddtrace/tracer test suite.

Since this test does not assert anything about the core tracer, move it
to the package that seems to be most appropriate.