chore(ddtrace/tracer): Fix two issues in TestPushTrace
Issue 1: Call tracer.Stop() instead of tracer.statsd.Close() to pass the
new goroutine leak detection when a real agent is running.

Issue 2: The test was depending on a previous test executing that
modified the logger level to debug. This means the test always failed
when running it in isolation.

TODO: We need to enable -shuffle for all our tests. IIRC we discussed
this in the guild before. I'll dig up the notes and start a discussion.