fix(profiler): don't double compress non-delta profiles (#3918)
Our non-delta heap, block, and mutex profiles are getting
double-compressed. That is, they are gzip-compressed, and if you
decompress them then you get another gzip-compressed blob, which is in
turn a compressed pprof file.

This is due to a bug introduced by #3529. In that PR we reworked our
compression logic in order to support re-compression (into zstd in
particular). We incorrectly decided whether a given profile type is a
delta profile by just checking whether there are delta values for the
profile type, without checking whether delta profiling is actually enabled.
As a result, when delta profiling is disabled we use the delta profling
enabled logic, where we assume the input data is uncomprossed and then
gzip-compress it.