feat(internal/appsec): activation metric and config key (#3959)
This PR adds the `DD_APPSEC_ENABLED` configuration that was supposed to already be there but was removed by stableconfig APM work. It also adds the `appsec.enabled` telemetyr metric to also report the value of appsec being enabled or not. This PR ALSO fixes a bug that stableconfig introduced by returning an error even when the env var is not set which made us branch out of appsec.Start too early. And since telemetry logs are dropped backend side we could not have find out earlier.

Signed-off-by: Eliott Bouhana <eliott.bouhana@datadoghq.com>