chore(ddtrace/tracer): fix flaky TestHealthMetricsRaceCondition
Switch wg.Wait() and flush(5), otherwise we might flush before the wait
group is done.

Stop relying on the total number of reported statsd metrics as an
indicator for the statsd client having received the metrics we are
looking for. In my testing, I've seen the client receive 200+ metrics in
this test (should it?), so performing our check after only 10 metrics
have reported seemed brittle.

Once we are on go1.25, we should start using the synctest package for
this test. The changes in this patch are expected to fix the flakiness
in a practical sense. But scheduling delays of > 5s will still cause
this test to flake.

[1] https://tip.golang.org/doc/go1.25#new-testingsynctest-package