fix(instrumentation): restore deleted net/http/pattern package
The instrumentation/net/http/pattern package was moved by #3853.  The
contribu/net/http/v2 module depends on that package at version v2.2.3.
As a result, that verison of contrib/net/http/v2 won't be able to use
newer versions of the main module. This is a breaking change. This
commit restores the deleted package. While the current
contrib/net/http/v2 package doesn't use it any more, this maintains
compatibility with previous versions.