fix(instrumentation): restore deleted net/http/pattern package (#4017 backport)
The instrumentation/net/http/pattern package was moved by #3853. The
contrib/net/http/v2 module depends on that package at version v2.2.3. As a
result, that verison of contrib/net/http/v2 won't be able to use newer
versions of the main module. This is a breaking change. This commit restores the
deleted package. While the current contrib/net/http/v2 package doesn't use it
any more, this maintains compatibility with previous versions.

In this backport, we don't have the NewSafeError functionality for
telemetry logging. Rather than bring it back to this branch, this commit
just uses the `%T` formatter to get roughly the same functionality.

Backports #4017