chore(telemetry): tighten scope that ticker func lock is held
The telemetry client flush method calls registered "ticker" functions.
These functions are in a list, which is guarded by a lock as it can be
modified concurrently. The flush method takes the lock and defers
unlocking it in a new scope (curly braces). However, defer arranges for
functions to run at the end of the enclosing function call, not the end
of the enclosing scope. So we're holding the lock longer than we need
to. To achieve the desired effect, move the calls to the tickers to its
own function.