fix(ddtrace/tracer): fix TestSpanString deadlock
TestSpanString takes a read lock on a span's mutex before calling
span.String. However, span.String also takes a read lock. The docs
for sync.RWMutex explicitly state that recursive read-locking is
prohibited. It can lead to deadlocks. For example:

- The test goroutine takes a read lock on the span.
- The tracer background goroutine takes a write lock in, e.g.
  span.setProcessTags. This blocks because there is a read lock held, as
  expected.
- The test goroutine then tries to take a read lock again on the span.
  The sync.RWMutex implementation sees that there is a pending write
  lock and blocks.

In this scenario, the test goroutine won't unblock until the write lock
is taken and released, which can't happen until the test goroutine
releases its initial read lock. So the test gets stuck. Fix this by not
taking the read lock in the test.