fix(ddtrace/tracer): deflake and fix race in TestEnqueuedTracesHealthMetric
TestEnqueuedTracesHealthMetric has two issues:

- It is flaky. The test uses tg.Wait(assert, 1, 10*time.Second) to wait
  for the specific metric it's checking to arrive. But this will just
  wait for 1 metric to arrive for up to 10 seconds. It stops waiting as
  soon as _any_ metric arrives. If we log the tg.Count() map on failure
  we can see that a bunch of other metrics are in the map.
- It has a race. The tracesQueued field of the agentTraceWriter should
  be accessed atomically.

This commit fixes those two issues. First, it uses assert.Eventually to
wait for the specific metric, datadog.tracer.queue.enqueued.traces, to
hit the expected value. Then it does an atomic read of the tracesQueued
value. This commit also removes the assert.True(ok) for the
agentTraceWriter type assertion. We can just do the type assertion
without the "ok" and let the test panic if the traceWriter isn't the
expected type.

Before this change, the test failed 5 out of 100 runs on my laptop.
After, it passed 10000 runs.