feat(tracer): add support for multiconfig for APMTracing RC product
Teach the RC update handler to handle multiple config files and merge
them according to priority. Advertise the APM_TRACING_MULTICONFIG
capability to RC in order to receive both org-level and service-level
configs. This brings the Go tracer in line with Java, Python, DotNet
which have implemented this capability recently.

This will be useful for controlling the Dynamic Instrumentation product
at both the org and the service/env level.