feat(remoteconfig): add support to the RC client for unsubscribing from a product
Unsubscribing removes a previously registered callback from being
invoked when its specific product receives an update. When the last
subscription for a particular product is remove, remote config is not
polled for that product anymore.

This will be useful for remotely disabling Live Debugger on a tracer:
once disabled, the previously-established subscription to the respective
RC product needs to be terminated.