feat(tracer): add support for dynamically enabling and disabling Live Debugger
This patch makes the tracer respect Remote Config updates for enabling
and disabling Live Debugger / Dynamic Instrumentation according to the
LiveDebuggingEnabled field on config sent to the APM_TRACING RC product.
These updates override the default, which comes from the
DD_DYNAMIC_INSTRUMENTATION_ENABLED env var (defaulting to false).

As far as the tracer is concerned, Dynamic Instrumentation being on
means that the tracer should subscribe to the LIVE_DEBUGGING and
LIVE_DEBUGGING_SYMBOL_DB RC products in order to receive probes (the
configs carrying the probes are snooped on by the local system-probe,
which will actually install the probes).

This patch also adds a WithDynamicInstrumentationEnabled(bool) option
for creating the tracer, with the same effects as setting the
DD_DYNAMIC_INSTRUMENTATION_ENABLED env var.